Add statistics and handle multi-part extraction

+++ core/src/test/java/io/sailrocket/core/AsyncEnv.java
@@ -58,7 +58,7 @@ public abstract class AsyncEnv {
 
 
     public int requestCount(){
-        return sequenceStats.requestCount.intValue();
+        return sequenceStats.requestCount;
     }
 
     protected volatile int count;
